+++ packages/database/src/QueryStatements/DatetimeStatement.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Tempest\Database\QueryStatements;
 
+use InvalidArgumentException;
 use Tempest\Database\Config\DatabaseDialect;
 use Tempest\Database\QueryStatement;
 
@@ -13,21 +14,32 @@ final readonly class DatetimeStatement implements QueryStatement
         private string $name,
         private bool $nullable = false,
         private ?string $default = null,
+        private bool $current = false,
     ) {}
 
     public function compile(DatabaseDialect $dialect): string
     {
+        if ($this->default !== null && $this->current === true) {
+            throw new InvalidArgumentException("Cannot set both `default` and `current` for datetime column `{$this->name}`.");
+        }
+
+        $default = match (true) {
+            $this->current => 'CURRENT_TIMESTAMP',
+            $this->default !== null => "'{$this->default}'",
+            default => null,
+        };
+
         return match ($dialect) {
             DatabaseDialect::POSTGRESQL => sprintf(
                 '`%s` TIMESTAMP %s %s',
                 $this->name,
-                $this->default !== null ? "DEFAULT '{$this->default}'" : '',
+                $default !== null ? "DEFAULT {$default}" : '',
                 $this->nullable ? '' : 'NOT NULL',
             ),
             default => sprintf(
                 '`%s` DATETIME %s %s',
                 $this->name,
-                $this->default !== null ? "DEFAULT '{$this->default}'" : '',
+                $default !== null ? "DEFAULT {$default}" : '',
                 $this->nullable ? '' : 'NOT NULL',
             ),
         };