+++ packages/reflection/src/ClassReflector.php
@@ -19,6 +19,7 @@ final class ClassReflector implements Reflector
     private array $memoize = [];
 
     private readonly PHPReflectionClass $reflectionClass;
+    private string $name;
 
     /**
      * @param class-string<TClassName>|TClassName|PHPReflectionClass<TClassName> $reflectionClass
@@ -34,6 +35,7 @@ final class ClassReflector implements Reflector
         }
 
         $this->reflectionClass = $reflectionClass;
+        $this->name = $reflectionClass->getName();
     }
 
     public function getReflection(): PHPReflectionClass
@@ -104,7 +106,7 @@ final class ClassReflector implements Reflector
      */
     public function getName(): string
     {
-        return $this->reflectionClass->getName();
+        return $this->name;
     }
 
     public function getShortName(): string
@@ -162,12 +164,12 @@ final class ClassReflector implements Reflector
 
     public function is(string $className): bool
     {
-        return $this->getType()->matches($className);
+        return is_a($this->getName(), $className, allow_string: true);
     }
 
     public function implements(string $interface): bool
     {
-        return $this->isInstantiable() && $this->getType()->matches($interface);
+        return $this->isInstantiable() && is_a($this->getName(), $interface, allow_string: true);
     }
 
     private function memoize(string $key, Closure $closure): mixed
@@ -187,5 +189,6 @@ final class ClassReflector implements Reflector
     public function __unserialize(array $data): void
     {
         $this->reflectionClass = new PHPReflectionClass($data['name']);
+        $this->name = $data['name'];
     }
 }