+++ packages/support/src/Filesystem/functions.php
@@ -620,7 +620,7 @@ function read_symbolic_link(string $path): string
  */
 function normalize_path(string $path): ?string
 {
-    if (class_exists(\Phar::class) && \Phar::running(false) !== '' && str_starts_with($path, 'phar:')) {
+    if (str_starts_with($path, 'phar:') && class_exists(\Phar::class) && \Phar::running(false) !== '') {
         return $path;
     }